🔒️ Corrige CWE-862

+++ src/main/java/com/unisinos/owasptoptwo/service/AnuncioService.java
@@ -1,11 +1,12 @@
 package com.unisinos.owasptoptwo.service;
 
+import com.unisinos.owasptoptwo.config.CustomUserDetails;
 import com.unisinos.owasptoptwo.model.Anuncio;
 import com.unisinos.owasptoptwo.repository.AnuncioRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -25,6 +26,12 @@ public class AnuncioService {
 	public void update(Integer id, Anuncio anuncio) {
 		var anuncioDesatualizado = anuncioRepository.findById(id).orElseThrow();
 
+		var user = getUsuarioLogado();
+
+		if (anuncioDesatualizado.getDono().getId() != user.getId()) {
+			throw new RuntimeException("Usuário não autorizado a atualizar o anúncio.");
+		}
+
 		anuncioDesatualizado.setTitulo(anuncio.getTitulo());
 		anuncioDesatualizado.setDescricao(anuncio.getDescricao());
 		anuncioDesatualizado.setDono(anuncio.getDono());
@@ -35,7 +42,19 @@ public class AnuncioService {
 
 
 	public void delete(Integer id) {
+		var user = getUsuarioLogado();
+
+		var anuncio = anuncioRepository.findById(id).orElseThrow();
+
+		if (anuncio.getDono().getId() != user.getId()) {
+			throw new RuntimeException("Usuário não autorizado a atualizar o anúncio.");
+		}
+
 		anuncioRepository.deleteById(id);
+	}
+
+	private CustomUserDetails getUsuarioLogado() {
+		return (CustomUserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 	}
 
 }
